Fix signed overflow due to implicit integer conversion

+++ lib/source/ctr_mode.c
@@ -58,8 +58,8 @@ int tc_ctr_mode(uint8_t *out, unsigned int outlen, const uint8_t *in,
 	(void)_copy(nonce, sizeof(nonce), ctr, sizeof(nonce));
 
 	/* select the last 4 bytes of the nonce to be incremented */
-	block_num = (nonce[12] << 24) | (nonce[13] << 16) |
-		    (nonce[14] << 8) | (nonce[15]);
+	block_num = ((unsigned int)nonce[12] << 24) | ((unsigned int)nonce[13] << 16) |
+		    ((unsigned int)nonce[14] << 8) | ((unsigned int)nonce[15]);
 	for (i = 0; i < inlen; ++i) {
 		if ((i % (TC_AES_BLOCK_SIZE)) == 0) {
 			/* encrypt data using the current nonce */
